requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community==0.0.10
langchain-openai>=0.0.5
openai>=1.0.0

# Database and data processing
python-dotenv>=1.0.0
sqlite3
pandas>=2.0.0
numpy==1.24.3

# Web search APIs
tavily-python>=0.3.0
google-search-results>=2.4.0

# HuggingFace datasets
datasets==2.15.0
huggingface_hub==0.20.2

# Utilities
requests>=2.31.0
streamlit>=1.28.0
python-dotenv==1.0.0
requests==2.31.0

# Optional: Redis for caching (if needed)
redis==5.0.1
